Made significant changes to the file structure in the backend. Deleted unnecessary or redundant files and folders. Make significant changes to several files to correct any path issues.
