--- JukeboxJam/music/apps.py
@@ -1,4 +1,0 @@
-from django.apps import AppConfig
-
-class MusicConfig(AppConfig):
-    name = 'music'
