--- django-settings/settings.py
@@ -1,15 +1,0 @@
-INSTALLED_APPS = [
-    'django.contrib.admin',
-    'django.contrib.auth',
-    'django.contrib.contenttypes',
-    'django.contrib.sessions',
-    'django.contrib.messages',
-    'django.contrib.staticfiles',
-    'authentication',
-    'music',
-]
-
-AUTH_USER_MODEL = 'authentication.CustomUser'
-
-# this is for CONSOLE USE
-EMAIL_CONSOLE = 'django.core.mail.backends.console.EmailBackend'
