+++ manage.py
@@ -1,23 +1,24 @@
-# import the OS and system
+#!/usr/bin/env python3
+
+"""Django's command-line utility for administrative tasks."""
+
 import os
 import sys
 
-# just like a main in C++ 
-def main():
-        "Run tasks"
-        os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'jukeboxjam.settings')
 
-        try:
-                from django.core.management import execute_from_command_line
-        except ImportError as exc:
-                raise ImportError(
-                        # these strings were written by AI as a substitute for errors
+def main():
+    """Run administrative tasks."""
+    os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'backend.django_settings.settings')
+    try:
+        from django.core.management import execute_from_command_line
+    except ImportError as exc:
+        raise ImportError(
             "Couldn't import Django. Are you sure it's installed and "
             "available on your PYTHONPATH environment variable? Did you "
             "forget to activate a virtual environment?"
-                ) from exc
-        
-        execute_from_command_line(sys.argv)
+        ) from exc
+    execute_from_command_line(sys.argv)
 
-        if __name__ == '__main__':
-                main() 
+
+if __name__ == '__main__':
+    main()
