+++ music/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+class MusicConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'music'
